pyproject.toml
[project]
name = "byeias"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "mistralai (>=2.0.2,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "datasets (>=4.7.0,<5.0.0)",
    "pypdf (>=6.1.1,<7.0.0)",
    "nltk (>=3.9.2,<4.0.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "torch (>=2.10.0,<3.0.0)",
    "transformers (>=5.3.0,<6.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "sentencepiece (>=0.2.1,<0.3.0)",
    "tiktoken (>=0.12.0,<0.13.0)",
    "protobuf (>=7.34.0,<8.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pre-commit (>=4.5.1,<5.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "ruff (>=0.15.6,<0.16.0)",
    "uvicorn (>=0.41.0,<0.42.0)"
]
